Refactor and break up player.rs
